Migration to Java 21
Introduced module-info.java to enable Java module system, moved JavaFX Application logic from Main.java to new App.java, and updated pom.xml to use Java 21 and latest JavaFX, ControlsFX, and GitHub API dependencies. This refactor modernizes the project structure, improves maintainability, and prepares the application for future enhancements.
