+++ pom.xml
@@ -9,7 +9,7 @@
     <url>https://github.com/koder95/eMetrykant</url>
     <groupId>pl.koder95</groupId>
     <artifactId>eMetrykant</artifactId>
-    <version>0.4.4</version>
+    <version>0.5.0-SNAPSHOT</version>
     <scm>
         <url>https://github.com/koder95/eMetrykant</url>
         <tag>search, metric, church</tag>
@@ -29,22 +29,34 @@
     </licenses>
 
     <properties>
-        <maven.compiler.target>1.8</maven.compiler.target>
-        <maven.compiler.source>1.8</maven.compiler.source>
+        <maven.compiler.target>21</maven.compiler.target>
+        <maven.compiler.source>21</maven.compiler.source>
         <project.build.sourceEncoding>UTF-8</project.build.sourceEncoding>
         <libs>${project.build.directory}/lib</libs>
+        <javafx.version>21.0.8</javafx.version>
+        <controlsfx.version>11.2.2</controlsfx.version>
     </properties>
 
     <dependencies>
         <dependency>
+            <groupId>org.openjfx</groupId>
+            <artifactId>javafx-fxml</artifactId>
+            <version>${javafx.version}</version>
+        </dependency>
+        <dependency>
+            <groupId>org.openjfx</groupId>
+            <artifactId>javafx-controls</artifactId>
+            <version>${javafx.version}</version>
+        </dependency>
+        <dependency>
             <groupId>org.controlsfx</groupId>
             <artifactId>controlsfx</artifactId>
-            <version>8.40.18</version>
+            <version>${controlsfx.version}</version>
         </dependency>
         <dependency>
             <groupId>org.kohsuke</groupId>
             <artifactId>github-api</artifactId>
-            <version>1.326</version>
+            <version>1.330</version>
         </dependency>
     </dependencies>
