+++ src/main/java/module-info.java
@@ -1,0 +1,13 @@
+module eme {
+    requires java.xml;
+    requires javafx.base;
+    requires javafx.controls;
+    requires javafx.fxml;
+    requires javafx.graphics;
+    requires org.kohsuke.github.api;
+    requires org.controlsfx.controls;
+
+    exports pl.koder95.eme;
+    opens pl.koder95.eme;
+    opens pl.koder95.eme.fx;
+}
